+++ docs/src/conf.py
@@ -3,8 +3,8 @@
 project = "Quadrant Todo"
 author = "Karl Han"
 copyright = "2026, Karl Han"
-version = "1.2"
-release = "1.2.0"
+version = "1.3"
+release = "1.3.0"
 
 extensions = []